Sync Switcheo Wallet With Attached NEO Wallet
The total coins shown in the Switcheo wallet balance often does not match the amount in the NEO wallet.  Sometimes it sync with time, sometimes it does not.   Yesterday the Neo wallet showed a lower amount of GAS, today it showed the correct number of Gas while the on both days the switcheo screen for account was incorrect.